updating models to standard base/internal/external format
